Project Files
- Added all project files.

One pdf is a documentation of all the tasks neccessary for the project to be considered complete.

The other pdf is command examples, to help students get the hang of SQL more easily.

The .csv files are files provided from the lecturer that hold all the data that our database needs to hold.